pyproject.toml
[tool.poetry]
name = "snovault"
version = "2.0.0"
description = "Storage"
authors = ["William Ronchetti <william_ronchetti@hms.harvard.edu>"]
license = "MIT"
readme = "README.rst"
homepage = "https://github.com/4dn-dcic/snovault"
repository = "https://github.com/4dn-dcic/snovault"
packages = [
  { include="snovault", from="." }
]
classifiers = [
    # How mature is this project? Common values are
    #   3 - Alpha
    #   4 - Beta
    #   5 - Production/Stable
    'Development Status :: 4 - Beta',

    # Indicate who your project is intended for
    'Intended Audience :: Developers',
    'Intended Audience :: Science/Research',
    'Framework :: Pyramid',


    # Pick your license as you wish (should match "license" above)
    'License :: OSI Approved :: MIT License',
    'Topic :: Database :: Database Engines/Servers',

    # Specify the Python versions you support here. In particular, ensure
    # that you indicate whether you support Python 2, Python 3 or both.
    'Programming Language :: Python :: 2.7',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.4'
]

[tool.poetry.dependencies]
python = "^3.4"
aws-requests-auth = "0.4.1"
awscli = "1.15.42"
bcrypt = "3.1.0"
beautifulsoup4 = "4.4.1"
boto3 = "1.7.42"
botocore = "1.10.42"
certifi = "2018.11.29"
cffi = "1.7.0"
chardet = "3.0.4"
colorama = "0.3.7"
coverage = "4.1"
cryptacular = "1.4.1"
dcicutils = "0.8.5"
docutils = "0.12"
elasticsearch = "5.5.3"
elasticsearch-dsl = "5.3.0"
flaky = "3.6.1"
future = "0.15.2"
futures = "3.1.1"
html5lib = "1.0b8"
humanfriendly = "1.44.7"
hupper = "1.5"
idna = "2.7"
isodate = "0.5.4"
jmespath = "0.9.0"
jsonschema-serialize-fork = "2.1.1"
keepalive = "0.5"
loremipsum = "1.0.5"
MarkupSafe = "0.23"
# # mr.developer = "1.38"
# mr.developer = ""
netaddr = "0.7.18"
passlib = "1.6.5"
PasteDeploy = "1.5.2"
pbkdf2 = "1.3"
Pillow = "3.3.0"
plaster = "1.0"
plaster-pastedeploy = "0.6"
psutil = "4.3.0"
psycopg2 = "2.7.3"
py = "1.4.31"
pyasn1 = "0.1.9"
PyBrowserID = "0.11.0"
pycparser = "2.14"
pyparsing = "2.2.0"
pyramid = "1.10.4"
pyramid-localroles = "0.1"
pyramid-multiauth = "0.8.0"
pyramid-retry = "1.0"
pyramid-tm = "2.2.1"
pyramid-translogger = "0.1"
pytest = "2.9.1"
pytest-cov = "2.2.1"
pytest-exact-fixtures = "0.1"
pytest-instafail = "0.3.0"
pytest-mock = "0.11.0"
pytest-runner = "4.2"
pytest-timeout = "1.0.0"
python-dateutil = "2.5.3"
python-magic = "0.4.11"
pytz = "2016.4"
PyYAML = "3.12"
rdflib = "4.2.2"
rdflib-jsonld = "0.4.0"
requests = "2.20.0"
rfc3987 = "1.3.6"
rsa = "3.4.2"
rutter = "0.2"
s3transfer = "0.1.13"
setuptools = "^36.6.0"
simplejson = "3.8.2"
six = "1.14.0"
SPARQLWrapper = "1.7.6"
SQLAlchemy = "1.2.16"
strict-rfc3339 = "0.7"
structlog = "18.1.0"
subprocess-middleware = "0.3"
transaction = "2.4.0"
translationstring = "1.3"
urllib3 = "1.23"
venusian = "1.2.0"
waitress = "1.2.0"
WebOb = "1.8.5"
WebTest = "2.0.21"
WSGIProxy2 = "0.4.2"
xlrd = "1.0.0"
# zc.buildout = "^2.13.2"
# zc.recipe.egg = "^2.0.5"
zope-deprecation = "4.4.0"
zope-interface = "4.6.0"
zope-sqlalchemy = "1.2"

[tool.poetry.dev-dependencies]
# pytest = "4.5.0"
pytest = ""
# pytest-cov = "2.7.1"
pytest-cov = "2.2.1"
# pytest-mock = "1.10.4"
pytest-mock = ""
# pytest-runner = "5.1"
pytest-runner = ""
flaky = "3.6.1"
# flake8 = "3.7.8"
flake8 = ""
# coverage = "4.5.4"
coverage = ""
responses = "^0"

[tool.poetry.scripts]
wipe-test-indices = "snovault.commands.wipe_test_indices:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "^poetry.masonry.api"
